Move code to lib and add a test
